Manage gamma correction panel
My problem: when chosing "manual", the "Opt." button was off-enabled, so no access to loading gamma factors from files, but on the other hand, it is weird to have access to photobleaching-based parameters when using manual.
My solution: change the pushbutton to "Load" when chosing manual and to "Opt." when choosing photobleaching-based.

For this:
- the function "pushbutton_loadGamma_Callback" was moved from "gammaOpt.m" to a separate file in order to allow calls from outside.
- the pushbutton calls "gammaOpt.m" when in photobleaching-based mode or "pushbutton_loadGamma_Callback.m" when in manual mode (see MASH.m)
- factor correction panel is updaed accordingly (see ud_cross.m)

Off-topic: in MASH.m;
- was also added the function edit_TP_states_param3_Callback that was missing (it must have been delected somehow)
- find states parameters were wrongly saved (using prm, the "used", instead of curr, the "to use" and the one display)