Modify third argument of callbacks
When the GUI is built programmatically, callback functions take the handle to MASH figure as the last input argument (and not the handle structure h as when the GUI is built with MATLAB's "guide" tool).
Here I modified the last input of each callback and adapted the content accordignly.

In addition, I modified the name of the kinetic rate edit fields so the edit field "edit_12" shows the value of rate k12 and not k21 as previously.

Finally, I have created an empty callback function for the "defocus" option in simulation's setup parameters, and a callback function for trace processing's top-plot popmenu that magically disappeared (!!!)